refactor: Move calls to the weather API to a new file
To declutter the main server.js file the calls to the weather API were
factored out into its own module.
