+++ src/server/server.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const bodyParser = require('body-parser');
 const cors = require('cors');
-const axios = require('axios');
 const dotenv = require('dotenv');
 dotenv.config();
 const { DynamicEntryPlugin } = require('webpack');
@@ -12,9 +11,10 @@ const geonamesAPI = require('./geonames-api');
 geonamesAPI.countryInfoService = countryInfoService;
 const pictureAPI = require('./picture-api');
 const pixabay = pictureAPI.createAPI(cacheService);
+const weatherAPI = require('./weather-api');
+weatherAPI.countryInfoService = countryInfoService;
 /* Global Variables */
-const weatherbitBaseURL = 'https://api.weatherbit.io/v2.0/';
-const weatherbitApiKey = process.env.WEATHERBIT_API_KEY;
+
 // Setup empty JS object to act as endpoint for all routes
 projectData = {};
 
@@ -55,141 +55,13 @@ app.post('/addData', (request, response) => {
     response.send(projectData);
 });
 
-app.get('/currentWeather', (request, response) => {
-    const query = getWeatherbitQuery(request);
+app.get('/currentWeather', weatherAPI.fetchCurrentWeather);
 
-    if (query) {
-        const url = weatherbitBaseURL + 'current?' + query;
-        axios
-            .get(url)
-            .then((weatherbitResponse) => {
-                if (weatherbitResponse.data.count > 0) {
-                    const weatherReport = weatherbitResponse.data.data[0];
-                    response.send({
-                        city: weatherReport.city_name,
-                        province: countryInfoService.getStateName(
-                            weatherReport.country_code,
-                            weatherReport.state_code
-                        ),
-                        country: countryInfoService.getCountryName(
-                            weatherReport.country_code
-                        ),
-                        code: weatherReport.weather.code,
-                        description: weatherReport.weather.description,
-                        temperature: weatherReport.temp,
-                        feelsLike: weatherReport.app_temp,
-                        windSpeed: weatherReport.wind_spd,
-                        windDirectionInDegrees: weatherReport.wind_dir,
-                        windDirectionAsText: weatherReport.wind_cdir,
-                    });
-                } else {
-                    response.send({
-                        error: 'No results were found for this location',
-                    });
-                }
-            })
-            .catch((error) => {
-                console.log(error);
-                response.send({
-                    error:
-                        'Something went wrong while retrieving the weather report',
-                });
-            });
-    } else {
-        response.send({
-            error: 'Not enough data to identify the desired location',
-        });
-    }
-});
-
-app.get('/forecast', (request, response) => {
-    const query = getWeatherbitQuery(request);
-
-    if (query) {
-        const url = weatherbitBaseURL + 'forecast/daily?' + query;
-        axios
-            .get(url)
-            .then((weatherbitResponse) => {
-                if (weatherbitResponse.data) {
-                    const weatherReport = weatherbitResponse.data;
-                    const dailyForecasts = weatherReport.data.map((report) => {
-                        return {
-                            date: report.valid_date,
-                            code: report.weather.code,
-                            description: report.weather.description,
-                            windSpeed: report.wind_spd,
-                            windDirectionInDegrees: report.wind_dir,
-                            windDirectionAsText: report.wind_cdir,
-                            maxTemperature: report.max_temp,
-                            minTemperature: report.min_temp,
-                            maxfeelsLike: report.app_max_temp,
-                            minfeelsLike: report.app_min_temp,
-                        };
-                    });
-                    response.send({
-                        city: weatherReport.city_name,
-                        province: countryInfoService.getStateName(
-                            weatherReport.country_code,
-                            weatherReport.state_code
-                        ),
-                        country: countryInfoService.getCountryName(
-                            weatherReport.country_code
-                        ),
-                        dailyForecasts: dailyForecasts,
-                    });
-                } else {
-                    response.send({
-                        error: 'No results were found for this location',
-                    });
-                }
-            })
-            .catch((error) => {
-                console.log(error);
-                response.send({
-                    error:
-                        'Something went wrong while retrieving the weather report',
-                });
-            });
-    } else {
-        response.send({
-            error: 'Not enough data to identify the desired location',
-        });
-    }
-});
+app.get('/forecast', weatherAPI.fetchWeatherForecast);
 
 app.get('/pictures', pixabay.retrievePicture);
 
 app.get('/listCities', geonamesAPI.fetchCityList);
-
-function getWeatherbitQuery(request) {
-    let query = '';
-    const latitude = request.query.latitude;
-    const longitude = request.query.longitude;
-    if (latitude && longitude) {
-        query += `lat=${encodeURIComponent(latitude)}&lon=${encodeURIComponent(
-            longitude
-        )}`;
-    } else {
-        const city = request.query.city;
-        const province = request.query.province;
-        const country = request.query.country;
-        if (city) {
-            query += `city=${encodeURIComponent(city)}${
-                province ? ',' + encodeURIComponent(province) : ''
-            }${country ? '&country=' + encodeURIComponent(country) : ''}`;
-        }
-    }
-    if (query) {
-        const timeToTrip = request.query.timeToTrip;
-        const duration = request.query.duration;
-        if (timeToTrip && duration) {
-            const days = parseInt(timeToTrip, 10) + parseInt(duration, 10);
-            query += `&days=${encodeURIComponent(days)}`;
-        }
-        query += `&key=${weatherbitApiKey}`;
-    }
-    return query;
-}
 
 // Gracefully close the system and save cache when termination signal is received
 // See https://expressjs.com/en/advanced/healthcheck-graceful-shutdown.html
